moves byline fixture data
