--- src/app/pages/ArticlePage/Byline/fixture/index.js
@@ -1,904 +1,0 @@
-export const bylineWithNoRole = [
-  {
-    type: 'contributor',
-    model: {
-      topicId: '',
-      topicUrl: '/news/topics/c8qx38nq177t',
-      blocks: [
-        {
-          type: 'name',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'Single Byline (all values)',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'Single Byline (all values)',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'link',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'test',
-                        blocks: [
-                          {
-                            type: 'urlLink',
-                            model: {
-                              text: 'test',
-                              locator: 'https://twitter.com/test',
-                              blocks: [
-                                {
-                                  type: 'fragment',
-                                  model: {
-                                    text: 'test',
-                                    attributes: [],
-                                  },
-                                },
-                              ],
-                              isExternal: true,
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'location',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'London',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'London',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'images',
-          model: {
-            blocks: [
-              {
-                type: 'image',
-                model: {
-                  blocks: [
-                    {
-                      type: 'rawImage',
-                      model: {
-                        width: 640,
-                        height: 562,
-                        locator:
-                          'f974/live/36226e20-94aa-11ec-9acc-37a09ce5ea88.png',
-                        originCode: 'cpsprodpb',
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-      ],
-    },
-  },
-];
-
-export const bylineWithNoAuthor = [
-  {
-    type: 'contributor',
-    model: {
-      topicId: '',
-      topicUrl: '',
-      blocks: [
-        {
-          type: 'role',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'Test',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'Test',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'link',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'test',
-                        blocks: [
-                          {
-                            type: 'urlLink',
-                            model: {
-                              text: 'test',
-                              locator: 'https://twitter.com/test',
-                              blocks: [
-                                {
-                                  type: 'fragment',
-                                  model: {
-                                    text: 'test',
-                                    attributes: [],
-                                  },
-                                },
-                              ],
-                              isExternal: true,
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'location',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'London',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'London',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'images',
-          model: {
-            blocks: [
-              {
-                type: 'image',
-                model: {
-                  blocks: [
-                    {
-                      type: 'rawImage',
-                      model: {
-                        width: 640,
-                        height: 562,
-                        locator:
-                          'f974/live/36226e20-94aa-11ec-9acc-37a09ce5ea88.png',
-                        originCode: 'cpsprodpb',
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-      ],
-    },
-  },
-];
-
-export const bylineWithNameAndRole = [
-  {
-    type: 'contributor',
-    model: {
-      topicId: '',
-      topicUrl: '',
-      blocks: [
-        {
-          type: 'name',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'Single Byline (all values)',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'Single Byline (all values)',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'role',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'Test',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'Test',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-      ],
-    },
-  },
-];
-
-export const bylineWithLink = [
-  {
-    type: 'contributor',
-    model: {
-      topicId: '',
-      topicUrl: '/news/topics/c8qx38nq177t',
-      blocks: [
-        {
-          type: 'name',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'Single Byline (all values)',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'Single Byline (all values)',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'role',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'Test',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'Test',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'link',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'test',
-                        blocks: [
-                          {
-                            type: 'urlLink',
-                            model: {
-                              text: 'test',
-                              locator: 'https://twitter.com/test',
-                              blocks: [
-                                {
-                                  type: 'fragment',
-                                  model: {
-                                    text: 'test',
-                                    attributes: [],
-                                  },
-                                },
-                              ],
-                              isExternal: true,
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-      ],
-    },
-  },
-];
-
-export const bylineWithLinkAndLocation = [
-  {
-    type: 'contributor',
-    model: {
-      topicId: '',
-      topicUrl: '',
-      blocks: [
-        {
-          type: 'name',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'Single Byline (all values)',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'Single Byline (all values)',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'role',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'Test',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'Test',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'link',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'test',
-                        blocks: [
-                          {
-                            type: 'urlLink',
-                            model: {
-                              text: 'test',
-                              locator: 'https://twitter.com/test',
-                              blocks: [
-                                {
-                                  type: 'fragment',
-                                  model: {
-                                    text: 'test',
-                                    attributes: [],
-                                  },
-                                },
-                              ],
-                              isExternal: true,
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'location',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'London',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'London',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-      ],
-    },
-  },
-];
-
-export const bylineWithNonPngPhoto = [
-  {
-    type: 'contributor',
-    model: {
-      topicId: '',
-      topicUrl: '',
-      blocks: [
-        {
-          type: 'name',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'Clark Kent',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'Clark Kent',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'role',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'Journalist',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'Journalist',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'link',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'superman',
-                        blocks: [
-                          {
-                            type: 'urlLink',
-                            model: {
-                              text: 'superman',
-                              locator: 'https://twitter.com/Superman',
-                              blocks: [
-                                {
-                                  type: 'fragment',
-                                  model: {
-                                    text: 'superman',
-                                    attributes: [],
-                                  },
-                                },
-                              ],
-                              isExternal: true,
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'location',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'Metropolis, US',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'Metropolis, US',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'images',
-          model: {
-            blocks: [
-              {
-                type: 'image',
-                model: {
-                  blocks: [
-                    {
-                      type: 'rawImage',
-                      model: {
-                        width: 640,
-                        height: 562,
-                        locator:
-                          'https://scarletjourney.rutgers.edu/crm/wp-content/uploads/sites/393/2017/12/christopher-reeve-superman.jpg',
-                        originCode: 'cpsdevpb',
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-      ],
-    },
-  },
-];
-
-export const bylineWithPngPhoto = [
-  {
-    type: 'contributor',
-    model: {
-      blocks: [
-        {
-          type: 'name',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'Mayeni Jones',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'Mayeni Jones',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'role',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'Journalist',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'Journalist',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'link',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'MayeniJones',
-                        blocks: [
-                          {
-                            type: 'urlLink',
-                            model: {
-                              text: 'MayeniJones',
-                              locator: 'https://twitter.com/MayeniJones',
-                              blocks: [
-                                {
-                                  type: 'fragment',
-                                  model: {
-                                    text: 'MayeniJones',
-                                    attributes: [],
-                                  },
-                                },
-                              ],
-                              isExternal: true,
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'location',
-          model: {
-            blocks: [
-              {
-                type: 'text',
-                model: {
-                  blocks: [
-                    {
-                      type: 'paragraph',
-                      model: {
-                        text: 'Lagos, Nigeria',
-                        blocks: [
-                          {
-                            type: 'fragment',
-                            model: {
-                              text: 'Lagos, Nigeria',
-                              attributes: [],
-                            },
-                          },
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-        {
-          type: 'images',
-          model: {
-            blocks: [
-              {
-                type: 'image',
-                model: {
-                  blocks: [
-                    {
-                      type: 'rawImage',
-                      model: {
-                        width: 640,
-                        height: 562,
-                        locator:
-                          'f974/live/36226e20-94aa-11ec-9acc-37a09ce5ea88.png',
-                        originCode: 'cpsprodpb',
-                      },
-                    },
-                  ],
-                },
-              },
-            ],
-          },
-        },
-      ],
-    },
-  },
-];
